Module kcore::dev

Device driver.

The design should be applicable for such use cases:

Env server

Tempdir and tempfile: env,

Two-way pipe: stdin/stdout, network

Event listener on file: Epoll

how to retrieve directory that posesses this file and then call its wstat.

Donot support atomic renaming due to portable reason, use special ctrl file instead.

Donot support link/unlink, use bind and mount instead. Since they provide more abstraction and behave more generally and more user-friendly.

Donot deal with access control, just r/w permission information.

Permission in file system?

Use namespace instead?

Traits

Device
Common trait of a device driver.