Struct kcore::vm::AddressSpace

pub struct AddressSpace<P: PageTable> {
    pub pgdir: P,
    // some fields omitted
}

Address space manager.

It manages the address space by a page table and segments. The page table describes the information of mapping from virtual address to physical address, while each segment describe one contiguous virtual memory with corresponding physical pages.

Memory management within the manageer is also divided into two parts. Page table handles pages used to build the table, while segments handles the allocation and deallocation of the mapped physical pages.

It's necessary to divide the function of address space into these two parts when implementing shared memory. For example, when the page table is freed, some shared physical pages should not be freed, thus, such information should be maintained outside the page table.

Fields
pgdir: P

Page table that contains mapping information.

Implementations

impl<P: PageTable> AddressSpace<P>

pub fn new() -> Result<Self, Error>

Create a new address space.

pub fn attach(
&mut self,
range: Range<usize>,
shared: bool,
overwrite: bool
) -> Result<bool, Error>

Add an non-empty segment.

Segments are not allowed to be overlap with each other. But can optionaly overwrite old segments.

Returns true if added. The overwriting procedure is NOT atomic, i.e., if overwriting failed, the overwritten segment will be dropped. But normal(non-overwrite) procedure is atomic.

pub fn detach(&mut self, saddr: usize) -> Result<bool, Error>

Remove a segment.

Returns true if removed.

pub fn share(&mut self, saddr: usize) -> Result<bool, Error>

Mark the segment as shared.

The target segment to be shared is any segment that contains saddr. Since segments do not overlap, it is unique. When forking, shared segments will be shared with same physical pages, while local segments will be copied with a new page of same content.

Return false if no such segment. Otherwise, return true.

pub fn segbrk(&mut self, saddr: usize, addr: usize) -> Result<bool, Error>

Change the segment from [old_start, old_end) to [old_start, addr).

addr should be greater than old_start to prevent creating empty segments. addr is required to be page-aligned. If addr > old_end, add new mappings and physical pages, else remove unnecessary mappings and physical pages. Shared segments cannot be changed.

This function is atomic. Return true if changed. Otherwise return false.

pub fn fork(&self) -> Result<Self, Error>

Fork address space and allocate new physical pages only for local segments.