Trait kcore::vm::PageTable

pub trait PageTable: Sized {
    const PG_LAYOUT: Layout;
    const PGSIZE: usize;
    const USERTOP: usize;

    fn new() -> Result<Self, Error>;

    fn switch(&self);

    fn map(&mut self, va: usize, pg: &Page<Self>) -> Result<(), Error>;

    fn unmap(&mut self, va: usize, len: usize);
}

Page table.

Associated Constants

const PG_LAYOUT: Layout

Layout used to alloc and dealloc physical pages from global allocator.

const PGSIZE: usize

Page size derived from [PG_LAYOUT].

const USERTOP: usize

Request map to the page table will be always lower than USERTOP. This must be page aligned.

Loading content...
Required methods

fn new() -> Result<Self, Error>

Create a new page table.

fn switch(&self)

Switch to this page table.

SAFETY: All pages mapped before switching should be accessiable after that.

That means the implementation can batch the map and unmap requests.

fn map(&mut self, va: usize, pg: &Page<Self>) -> Result<(), Error>

Map virtual address [va, va+pg.size()) to physical page pg.

It is required to be atomic, i.e., undo the mapping if failed. Caller guarantees that pages in [va, va+pg.size()) are unmapped.

fn unmap(&mut self, va: usize, len: usize)

Unmap the pages in [va, va+len) in this page table.

Caller guarantees that [va, va+len) has been mapped before.

Loading content...