Trait kcore::dev::Device

pub trait Device {
    fn typeid(&self) -> usize;

    fn reset(&mut self);

    fn init(&mut self);

    fn shutdown(&mut self);

    fn power(&mut self, on: bool);

#[must_use]    fn attach<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        name: &'life1 [u8]
    ) -> Result<Pin<Box<dyn Future<Output = Result<Arc<Chan>, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn walk<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan, 
        name: &'life2 [u8]
    ) -> Result<Pin<Box<dyn Future<Output = Result<Option<Chan>, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn stat<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan
    ) -> Result<Pin<Box<dyn Future<Output = Result<Dirent, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn wstat<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan, 
        dirent: Dirent
    ) -> Result<Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn open<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan, 
        omode: u32
    ) -> Result<Pin<Box<dyn Future<Output = Result<Chan, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn create<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan, 
        name: &'life2 [u8], 
        omode: u32, 
        perm: u32
    ) -> Result<Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn remove<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan
    ) -> Result<Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn close<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan
    ) -> Result<Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn read<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan, 
        buf: &'life2 mut [u8], 
        off: usize
    ) -> Result<Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn write<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        c: &'life1 Chan, 
        buf: &'life2 [u8], 
        off: usize
    ) -> Result<Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>>, AllocError>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

Common trait of a device driver.

Required methods

fn typeid(&self) -> usize

Unique character representing this device, such as #s. Unique name of the device. Type determine which driver to used for this channel, analogous to UNIX's major number.

fn reset(&mut self)

fn init(&mut self)

fn shutdown(&mut self)

fn power(&mut self, on: bool)

#[must_use]fn attach<'life0, 'life1, 'async_trait>(
&'life0 mut self,
name: &'life1 [u8]
) -> Result<Pin<Box<dyn Future<Output = Result<Arc<Chan>, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

#[must_use]fn walk<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan,
name: &'life2 [u8]
) -> Result<Pin<Box<dyn Future<Output = Result<Option<Chan>, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Looks for the file name in the directory represented by a channel. FIXME: why different from plan9.

#[must_use]fn stat<'life0, 'life1, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan
) -> Result<Pin<Box<dyn Future<Output = Result<Dirent, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Inquires about the file attribute identified by a channel.

#[must_use]fn wstat<'life0, 'life1, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan,
dirent: Dirent
) -> Result<Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Change the file attribute identified by a channel.

#[must_use]fn open<'life0, 'life1, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan,
omode: u32
) -> Result<Pin<Box<dyn Future<Output = Result<Chan, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Open a file for reading or writing.

#[must_use]fn create<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan,
name: &'life2 [u8],
omode: u32,
perm: u32
) -> Result<Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Create a file for reading or writing.

#[must_use]fn remove<'life0, 'life1, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan
) -> Result<Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Remove a file.

#[must_use]fn close<'life0, 'life1, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan
) -> Result<Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Close a file.

#[must_use]fn read<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan,
buf: &'life2 mut [u8],
off: usize
) -> Result<Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Read data from a channel by offset and store to buffer. Return the number of bytes read.

#[must_use]fn write<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
c: &'life1 Chan,
buf: &'life2 [u8],
off: usize
) -> Result<Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>>, AllocError> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Write the data in buffer to a channel starting by specific offset. Return the number of bytes writed.

Loading content...