Trait kcore::vm::PageTable

pub trait PageTable: Sized {
    const USERTOP: usize;
    const PAGE_LAYOUT: Layout;
    const PAGE_SIZE: usize;

    fn new() -> Result<Self>;

    fn switch(&self);

    fn map(&mut self, va: usize, vpa: usize) -> Result<()>;

    fn unmap(&mut self, va: usize) -> bool;

    fn protect(&mut self, va: usize, read_only: bool);
}

Page table.

Associated Constants

const USERTOP: usize

Request map to the page table will be always lower than USERTOP. This must be page aligned.

const PAGE_LAYOUT: Layout

Layout used to alloc and dealloc physical pages from global allocator.

const PAGE_SIZE: usize

Page size derived from [PAGE_LAYOUT]. Donot overwrite this.

Required methods

fn new() -> Result<Self>

Create a new page table.

fn switch(&self)

Switch to this page table.

SAFETY: All pages mapped before switching should be accessiable after that.

That means the implementation can batch the map and unmap requests.

fn map(&mut self, va: usize, vpa: usize) -> Result<()>

Map virtual address [va, va+PAGE_SIZE) to physical page at vpa with write permission.

It’s guaranteed that va is page-aligned. Overwrite the old mapping if exists. The implementation is required to be atomic, i.e., undo the mapping if failed.

fn unmap(&mut self, va: usize) -> bool

Unmap the pages in [va, va+PAGE_SIZE) in this page table.

Returns false if there is no such page. It’s guaranteed that va is page-aligned.

fn protect(&mut self, va: usize, read_only: bool)

Protect the page in [va, va+PAGE_SIZE) as read-only or read-write.

It’s guaranteed that va is page-aligned.